CLOUD-11389 - Menu updates for User Settings
## Problem

Design wants the user settings menu to be groups by user types. This involves separating the users by their type and sorting them alphabetically

## Solution

Refactor the existing logic on how user items are built for the menu. Separate them and sort them.

They also asked if we could keep the accent colors on the icons. Each icon has a little indicator in front of the user and that is supposed to be it's own color to stand out. So I had to create a new color variable to match it

| | BEFORE | AFTER |
| ------ | ------ | ------ |
| DARK |    ![Screenshot_2023-11-09_at_2.23.53_PM](/uploads/8d5d99c771dff1ea9d97d7c1b9beedfa/Screenshot_2023-11-09_at_2.23.53_PM.png)    |  ![Screenshot_2023-11-09_at_2.21.56_PM](/uploads/8c28c9adf381d5038d440624c7a1cde7/Screenshot_2023-11-09_at_2.21.56_PM.png)     |
| LIGHT |    ![Screenshot_2023-11-09_at_2.24.28_PM](/uploads/070388f9fc2820db071e2a6325bf346b/Screenshot_2023-11-09_at_2.24.28_PM.png)    |  ![Screenshot_2023-11-09_at_2.22.16_PM](/uploads/df8139ff8b6dee12c4c26f3d52524674/Screenshot_2023-11-09_at_2.22.16_PM.png)       |