CLOUD-11729: Fixed logic for custom permissions banner
# Problem
The custom permission banner was showing for users without custom permissions. A new property, hasPermissions, was recently added to check for custom permissions. Prior to the implementation of the permission manager, an aggregated permission string was being created via the getPermissionsFromUserGroups method. When assigning the hasPermissions property, calling normalizePermissionString for user.permissions was causing the check for permissionString !== 'none' to evaluate inaccurately because other permissions (other than custom permissions) were being added to the string due to the use of the no longer needed method.

# Changes
Now, with the implementation of the permission manager, the getPermissionsFromUserGroups method is no longer needed.  Removed the use of normalizePermissionString (which added "non-custom" permissions), as well as getPermissionsFromUserGroups (and associated variables) in order to prevent the inaccurate evaluation of permissionString.