CLOUD-11638: Updates to appearance of groups on hover
# Problem
The newly added tooltip's position was incorrect (higher than on the spec) and the word "and" was missing before the last item on the groups list. Also, due to the hard coded viewHeight and viewLines for the multiLine ellipsis component the tooltip was always the same height, regardless of the number of lines, causing unwanted white space at the bottom of the tooltip when less than 6 lines.

# Changes
- Tooltip position: added a new position property to the tooltip component in order to address the top property.
- Add "and" to list: added a new listItems property to the language file that takes a list of items and a single item and then joins them with the string "and".
- Added a calculation for viewHeight and viewLines based on the length of the accessRole string.