CLOUD-11846: Fix Double scroll on Bookmarks page when zoomed in
# Problem
The Bookmarks page allows the user to scroll down to load more bookmarks. However, when you zoom in on the page, a second scroll bar will appear to allow the user to scroll on the page. Only the sidebar and bookmarks should be scrollable, not the page itself

# Changes
Set the Bookmark Sidebar's position to `relative`

# Before
![Screen_Shot_2023-11-16_at_1.15.09_PM](/uploads/34882a82d33972de337116f0abe2ab4d/Screen_Shot_2023-11-16_at_1.15.09_PM.png)

# After
![Screen_Shot_2023-11-16_at_1.14.57_PM](/uploads/a59bc9f084a46602c8e11a8422d0f299/Screen_Shot_2023-11-16_at_1.14.57_PM.png)