CLOUD-11941 - Fix bug switching cameras in settings
## Problem

When switching cameras the settings menu doesn't update the selected camera

## Solution

This was because we recently made a change (https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/5147) that altered the way `camera` was given as an input to the camera settings.

I think it's easier to manage for now to revert that part of the change. The main goal was to get a signal version of the cameraId so I kept that part. Ideally the camera thats an input would just be a signal itself and then we don't have to bother with that, but we will have to wait for Angular 18 to use that pattern I guess.

Also there is a refactor like 75% done with camera settings that removes most of this code anyway. Also also, there is a redesign coming that probably makes all of this worthless anyway :upside_down: 

### TLDR;
reverting a bit of a change that caused the menu to stop updating for camera settings.