CLOUD-11494: Add another function to get License Summary
# Problem
After activating a License key, the License summary would only _sometimes_ display because we have a decorator/memoize on `getLicenseSummaries()`. We need the License summary to display immediately after activation

# Changes
We'll add another function called `getLicenseSummariesOnActivation()`. This does the same exact thing as `getLicenseSummaries()`, except it doesn't have a decorator/memoize. We'll use this function only when activating a license.

# Notes
We _can_ just remove the `memoize` on `getLicenseSummaries()` and call it a day, however, this will mean more API calls when navigating to the License page and I don't think we'd want this outcome