CLOUD-11936 - Camera Icon in menu for Archived cameras fix
## Problem

The camera menu does not show the grey circle when a camera has an archive, but is NOT recording.

## Solution

1. The data that comes in has a dirtyId, but we were checking against a cleanId. So it would never match. This is so painful that its 2024 this is even possible :sob: 
2. We fetch the camera archive data, but it loads async and there is a race condition between it and building the menu.

We could solve this by waiting to build the whole menu after we load the archive data, but this makes the page load slower. Instead I pulled out the camera specific menu stuff into a function and we call that only after the archive data loads

| BEFORE | AFTER |
| ------ | ------ |
|    ![image](/uploads/afd0c01a54d7d4068969542150840521/image.png)    |    ![image](/uploads/c63bea33b4d8ab253b024b344f2facc9/image.png)    |