CLOUD-11925 - Added conditional logic for how editable groups are displayed
# Problem
If a user is a part of just one editable permission group the UI should be different from that of a user who is a part of multiple editable permission groups. However, they were being displayed in the same way.

# Solution
Added conditional logic to the template to show different UIs based on the number of permission groups.