CLOUD-11903, CLOUD-11817: Rebuild system menu on activeSystem info update
# Problem
Buggy behavior in header level 2 (system menu):

1. View tab would sometimes not show for a user with custom permissions (resource-specific permissions, rather than group permissions)

2. The menu sometimes wouldn't update when a system was selected

The root cause of these issues was that rebuilding the menu ran in an effect that would only re-run on an update to `activeSystem$$()` or `activeSystem$$().permissionManager.permissions$$()` . When `activeSystem$$` was first set, sometimes the cameras wouldn't be loaded yet, so resource-specific permissions could not evaluate properly (bug #1 above). When a user went back to view all systems and then selected the same system, there would not be a change to `activeSystem$$`, so the menu would not rebuild/get updated (bug #2).

# Solution
Update the system menu to rebuild on any update from the `activeSystem.infoSubject` observable. Also update the system name in the header logo area to update following the same logic, so that its loading is synchronized with the header tabs.