CLOUD-11635, CLOUD-11636: Standardize feature flag usage
# Problem

We currently have several different ways of using a feature flag in the frontend.

# Solution

Standardize feature flag usage so that there is just one consistent way, which is to reference `nxConfig` directly everywhere except for in Angular templates, where we would reference nxConfig assigned to `CONFIG` class member in the component.

```
class myComponent {
  CONFIG = nxConfig;

  if (nxConfig.featureFlags.myFlag) { ... }
}

<div *ngIf="CONFIG.featureFlags.myFlag; else XYZ"></div>
```

No longer will we reference feature flags using NxConfigService

```
class myComponent {
  CONFIG: IConfig;
  constructor(configService: NxConfigService) {
    this.CONFIG = configService.getConfig();
  }
}
```

or 

```
class myComponent {
  isEnabled = true;
  constructor(configService: NxConfigService) {
    this.isEnabled = this.configService.flagsEnabled(FeatureFlagStrings.readonlyAPIs)
  }
}
```

In addition to consistency, this would make it easier to migrate feature flags out of the app config if we decide to do so in the future. Standardizing the usage of config throughout the codebase is outside the scope of this ticket.



Additional related changes:
- Remove the `flagsEnabled()` method from the configService
- Simplify feature route guards by removing unneeded cookie override logic