Add ivpd test #8, step 1
This test opens ipvd, clicks on the filter button "Ptx", searches for a particular manufacturer and makes sure that manufacturer is listed in the table. The line
    ipvd_page.search_text(manufacturer)
requires a sleep to allow the page to reflect the query. I know that putting in sleeps is not ideal, in this case,ipvd_page.search_text() is code that is only used in the ipvd tests and I really don't see how the extra effort needed to avoid the sleep would be worth while.