CLOUD-11815, CLOUD-11876, CLOUD-11998: Auth fixes
# Problem 1 (CLOUD-11815)
TOS didn't work because account was none before registering

# Solution 1
Fetch the account model after the user has been registered in cloud portal.

# Problem 2 (CLOUD-11876)
If you try to use oauth in the desktop client the previous user prevents you from logging into a new user

# Solution 2
Update the logic for locking the sub header so that it's unlocked on login attempts, but locked on password confirmation ones

# Problem 3 (CLOUD-11998)
The text for session expired lacked context for why the user was logging in. Since technically their session was still good for non-owner actions.

# Solution 3
Update the text to be more generic so that the user knows they are logging in to apply changes. While we could make the messages more generic it will give us more to manage without any significant benefits for us or the user.