CLOUD-11998: Initialized Oauth to the email state
# Problem
The oauth dialog has a race condition in the production bundle where verifying the vms takes longer than expected. As a result the dialog isn't in the correct state when checking the email so the user is stuck on the email step.

# Solution
Initialize the oauth dialog in the email state. So that it can jump to the password state. If the redirect is deemed insecure behavior is unchanged.