CLOUD-12028: Fix live status indicator
# Changes
The logic for wether or not a video is paused is pretty flawed and tightly coupled.

Added a playerActive property to keep track of if the player is actually mounted.

Not ideal but it's a pretty safe change which is probably better than trying potentially
breaking other things by updating the behavior around playback.livePaused.

# Reason for the change
Live status indicator wasn't showing correctly for offline cameras