CLOUD-11745: removes empty space at bottom of permission groups tooltip
# Problem
Dynamically calculating the viewHeight to pass to the multiline-ellipsis component was inconsistently creating unnecessary padding at the bottom of the tooltip.

# Solution
Added a new boolean input, fitContent to the mle. If true, the height will be set to 'fit-content'; otherwise it will be set to viewHeight input. In the admin component, if the viewLines are less than 6, then fitContent is set to true.

*There is still an outstanding issue regarding attaching the ellipsis to the end of the string that will be addressed in a separate MR.