CLOUD-11905: Fix settings menu loading bug
# Problem

When accessing offline and 4.2 systems (unable to reproduce for 5.0.0.36634 system) in Cloud, the system settings menu loads indefinitely. The reason is that `.some()` is called on system cameras in `updateCameraSettingsMenu()` when cameras is still undefined, causing `updateMenu()` to error and exit early before the menu is finished building.

# Solution

Initialize `system.cameraManager.cameras` as an empty array rather than undefined.