CLOUD-11954: Add error message for Cloud to local system merge
# Problem

When attempting to merge a cloud system into a local system via WebAdmin, the merge API call fails (expected behavior). 

```
{
    "error": "4",
    "errorId": "forbidden",
    "errorString": "A non-Cloud and %1 system cannot be merged, provided that the non-Cloud system is specified as the System from which the System name and settings will be taken when merging."
}
```

However, an "unknown error" message is shown in the frontend.

# Solution

Check for this particular context in the frontend pre-merge check, before making the merge API call. If found, render a descriptive error message to the user.