CLOUD-12038 - Temp users group name is missing in settings
## Problem 

Temp Users group name was not being displayed in settings

## Solution

This is because it did not have the marker saying Temp Users are not editable. There could be some refactoring of this area, because now Temp users have logic in a few places in the User Settings. For now this seems okay since this is the end of Temp Users work for this release.

| Before | After |
| ------ | ------ |
|   ![image](/uploads/715f4b1b74069c384fbbd2abd60f57e7/image.png)     |    ![image](/uploads/44e7752d261fb7598807817068c5921a/image.png)    |