CLOUD-11989 - Fixing red circle in camera settings menu
## Problem

The SVG for scheduled recording is a red circle, and it should stay red in the menu even when hovered and selected. Currently when you hover and select that camera the circle will turn white

## Solution

This is caused by the menu component changing the color of the SVG dynamically. There already exists a pattern to stop this with the full filled in red circle SVG, and that is by adding `!important` directly to the SVG.

I know this pattern is bad and we shouldn't do it like this, but this involves a lot of changes to icons and to menu code. So I'm not gonna do that now. I'll cave and just add the important and move on :sob: 

| BEFORE | AFTER |
| ------ | ------ |
|    ![image](/uploads/7875fea912345ebaf59a393776c9b9df/image.png)    |    ![image](/uploads/7b2538449a602e4f9c0ea321973f67ac/image.png)    |