CLOUD-11742 CLOUD-11621 - advanced settings fix webadmin
## Problem

Advanced settings doesnt work on WebAdmin right now on 6.0

## Solution

The issue was that you couldn't navigate there due to some routing changes. I removed an extra route in the web admin routing table that was breaking it.

In short, there was a path match of `path: 'settings/advanced'` instead of letting `path: 'settings'` catch it. So then when it got to the `settings.module` routing it didn't catch it properly.

You had to go to `settings/advanced/advanced` to make it work