CLOUD-11982: Fix Webadmin errorStrings not displaying properly
# Problem
When trying to log into Webadmin Local and the browser's locale is set to another country/language that's not English, the login `errorStrings` are being returned in that language. Since it's not in English, we're unable to map the error to our list of errors

# Changes
We'll map the errors with `errorId` instead of `errorString`. `errorId`s are in "English"