Add vms_server 6.0 compatibility and new user navigation to 1 test
Added compatibility for 6.0 vms_server by refactoring small things.
Changed docker host ip to working one after Burbank outage.
Added the 'id' attribute to CloudAccount which is populated whenever
a CloudAccount is added to a cloud system. That id is returned by the
cloud as 'vmsUserId' and by the server as 'id'. Other ids are associated
with a cloud account, but they are not used in navigation and not shared
with the vms server.
Refactored user left_menu navigation for new system left menu. Tested on
12/12/2023 cloud-test instance. Made 1 test work with new navigation.
You may now set the vms version at the top of suite.py.