CLOUD-11745 - removes unwanted spaces in tooltip
# Problem
In the permission group's tooltip, depending on the length of the string being truncated, there was unwanted space between the end of the string and the ellipsis. Also, there was unnecessary padding at the bottom of the tooltip.

# Solution
Created a new component that uses -webkit-line-clamp to add an ellipsis to the end of truncated text. The property accepts a positive integer to determine the number of lines to show before truncating. This solution solves for both unwanted spaces, between the end of string and ellipsis, as well as at the bottom of the tooltip.