CLOUD-11520: Display "noServerUrl" error message on Merge
## Problem

On Webadmin, when you type in an invalid URL into the Merge input and the Mediaserver doesn't find a server, we're getting an "unexpected error" message.

# Changes

When the `info` response fails cause it couldn't find the server, it returns 0, so we'll display the "No Server Found" message when the `err.status === 0`.

Related to CLOUD-11520