CLOUD-12190: Removed userWithGroups feature flag from the frontend
# Problem
When the app is served in cloud mode it depends on the userWithGroups feature flag that was used for development.

# Solution
Remove the userWithGroups and depend solely on the system version now that users are stable.

(cherry picked from commit b7553dd3a140f7a67153b8dd94d35f10af86d5fd)