CLOUD-12086, CLOUD-12213: Made email not lock when in connect mode
# Problem
There was a change to lock changing emails for owner actions. As a consequence connect was affected.

# Solution
Make the email not locked when the clientType is connect. This lets you change users and brings back the back and forgot password buttons

(cherry picked from commit a906ceba377266fba048cf6d342c935cecc3896b)