CLOUD-12126, CLOUD-12143: Updated how sessions are managed in the frontend
The session service was tied into a lot of components and was primarily used for getting the current users email. In addition it also managed some other data by acting as an api for the local storage service.

Refactor the session service so that its primarily focused on managing the users session on the frontend. This includes when multiple tabs are opened in the browser.

- Migrated all uses of the loginStateSubject to the nxrg store using accountSelector.selectCurrentEmail
- Where applicable sessionService has been replaced with the localStorageService.
- Minor fixes where undefined objects were being referenced.
- Added a wrapper on dexie to prevent the random db from being delete in all browser tabs.
- Reduced dexies footprint within components.

(cherry picked from commit f7634abfd78c60e0a0cb257e8b93a769f20a3b24)