CLOUD-12126, CLOUD-12143: Updated how sessions are managed in the frontend
# Problem
The session service was tied into a lot of components and was primarily used for getting the current users email. In addition it also managed some other data by acting as an api for the local storage service.

# Solution
Refactor the session service so that its primarily focused on managing the users session on the frontend. This includes when multiple tabs are opened in the browser.

# Changes
- Migrated all uses of the loginStateSubject to the nxrg store using accountSelector.selectCurrentEmail
- Where applicable sessionService has been replaced with the localStorageService.
- Minor fixes where undefined objects were being referenced.
- Added a wrapper on dexie to prevent the random db from being delete in all browser tabs.
- Reduced dexies footprint within components.