CLOUD-11624, CLOUD-11659: Logout of all systems when changing users
When the user changes to a new one on cloud portal the previous cookies are kept giving the new user the same access level as the previous user.

Updated the session service so that it triggers all of the systems to logout before changing to the new user.

(cherry picked from commit 6a629460096287d6629dfeeca20751cd4c7fc361)