CLOUD-11854: Added a new middleware to prevent user from being able to login
# Problem
If there's a cloud outage sometimes users will try to log out and login again. When this happens the csrftoken is dropped but the sessionid stays behind. When cloud portal comes back online they try to login again but cant due to a missing csrf token.

# Solution
Rather than force users to clear their site data we try to remove the sessionid cookie when the csrftoken is missing. This prevents the csrf middleware from blocking the users attempts to login.