CLOUD-11891 - CLOUD-12080 - Motion Mask refactor
## Problem

After we refactored a lot of the camera settings, there were a bunch of Motion Sensitivity bugs. The Motion Sensitivity components were a bit tied to timings and state resets happening in certain orders. This worked for awhile, but as I tried to pull it into a reusable component it made it difficult to work with.

## Solution

I tried patching the bugs but eventually I decided to just refactor the component as best as I could to remove some of the timing/state issues. 

Feel free to ask questions if you want to know why I did something. This component is pretty complicated (with good reason. It's a complicated thing to make!) so I did my best to understand it and improve on it. There are still a few small things I would like to improve, but for now I have spent too much time and need to get it reviewed.

## What UX changed?

The motion mask stuff should be much quicker and more stable when changing cameras or rotation/aspect ratio.

## Known bugs

I think all of these are existing bugs..

* When you click reset after making changes, the black borders around the sections can mess up.
* When you Enable and Disable the motion detection section, sometimes the preview image doesn't show back up. You have to refresh the page
* When you click a number to set update the selected area, you can't click the same number multiple times.

(cherry picked from commit efd0679e83e7dc0ce0ab7b1970a967637c0c2bbd)