CLOUD-12194: Update Merge API to v3 and fix merging with unconfigured systems
# Problem

Merge was throwing a 422 error whenever we would try and merge with a system that is unconfigured. I believe the response here changed, so we just need to check if the system is new (unconfigured)

# Solution

We'll need to update the merge API to v3 so we can use the new functionality in the merge request for 6.0. Cloud merge uses CDB, meanwhile Local uses the latest version, so v1 will not be supported anymore

# Changes

* Add new checks to see if the system is new (unconfigured)
* Update `mergeSystems` to use v3. If we try to use v1, we'll throw an error. The API needs to be updated to v3 anyway, so we'll also need to move the function to the v3 file
* Add types for `mergeSystems`

Related to CLOUD-12194

(cherry picked from commit d18164eec5dfc3064145b387ec348785233adc2e)

# Conflicts:
#	front_end/libs/services/system-api.types/system-api.types.ts
#	front_end/libs/services/system-rest-api-v3.service.ts
#	front_end/libs/services/system-rest-api.service.ts