CLOUD-12224: Added a default value for deviceType on the view tab
# Problem
In 5.0 devices doesn't return the device type as a result it's undefined which causes all cameras to get filtered out.

# Solution
If the deviceType is missing we assume it's a camera and add it to the response for devices. Its sorta fixable but only for admins because it requires access to health metrics which does include a devices type.