CLOUD-12322: Fix check for no bookmarks
# Problem

When there are no bookmarks, the "No Results" placeholder is displaying instead of the "No Bookmarks" placeholder

# Changes

Add a more specific check for `creationCutOffTimeMS$` and see if it's `> 1.` If it's not `> 1`, then that means there are no New Bookmarks. We're already checking for `!bookmarks.length` in the parent container. If there are no bookmarks and there are no new bookmarks, then we display the "No Bookmarks" placeholder accordingly

Related to CLOUD-12322