CLOUD-12334: Updated the local-system-status-interceptor to update loginstate before reloading.
# Problem
Due to changes with the login state a browser can be authorized from a previous session. This triggers it to fetch for an account which causes other error handlers to be called incorrectly causes a logout loop.

# Solution
When the users local session is over set the loginState to unauthorized so that the loop will not start.