CLOUD-12334: Updated the local-system-status-interceptor to update loginstate before reloading.
Due to changes with the login state a browser can be authorized from a previous session. This triggers it to fetch for an account which causes other error handlers to be called incorrectly causes a logout loop.

When the users local session is over set the loginState to unauthorized so that the loop will not start.

(cherry picked from commit 5c1280c6621dd775219a0527e583e7145755db8e)