CLOUD-12339 - CLOUD-12094 - Camera Authorization bugs
1. The request to update the camera authorization was not working
2. When the authorization for a camera was updated, it was not refreshing the data in camera settings
3. Small bug clogging up console log for svg that doesnt exist

We need to go away from using the EC2 APIs and use REST. For this I created a new API handler just for updating the password.

We also need to update the camera settings component whenever the auth update request finishes. This is not a beautiful solution, but it's good enough for now because this component will be remade soon.

(cherry picked from commit 23af9809de28c22a8a2f411b490057ce7f8d4c91)