Add Test class and artifact saver
Add a method to chrome.py::ChromeBrowser to capture and save Chrome logs
and screenshots into the PythonRobot/artifacts folder. Files are named:
{test_case}_{timestamp}_chrome_logs.json (or {test_case}_{timestamp}_screenshot.png).

Add the ability to run the entire suite, catching exceptions along the way, printing
results and running all the tests in the suite.

Add results to console with pass/fail/skip, stack trace and durations.