CLOUD-11901: Fixed bug where cloud user wasn't editable
# Problem
If the user was recently added without reloading the user's fullName was missing. So when we patched the users update it tried to modify the users fullName which isn't allowed for cloud users as its controlled by cdb and not the system.

# Solution
When saving the user remove the fullName key to prevent the issue.

(cherry picked from commit 47fb12f54c740e4ca46e702ddbf352eec5d6f910)