CLOUD-12294: Update WebAdmin login error handling
# Problem

When submitting an invalid entry to the WebAdmin login form without changing any input values first (such as submitting immediately after a page refresh), no error message is shown to the user, although the inputs are highlighted red. Intended behavior is to show an error message.

The reason was that error messages in the UI were only being set in the login process, but when the submit process button was clicked when the form was untouched, the login process would not fire. So the form inputs would highlight red due to validation errors, but no error message would show.

# Solution

Add a click handler to the process button so that it shows an error message if the form is not valid, even if the login process does not fire.