CLOUD-11702: Routing to "No Cameras" page when no permission
# Problem
When a user with no permissions had access to a system with a camera, trying to route to the cameras tab with the direct link resulted in a blank page. This was due to the getCamera method not returning a value if there are no cameras to get.

# Solution
Updated patchCameraCompatibility method to include conditional logic that returns undefined if there is no camera. A user will now see a toast alert indicating that there's no camera for them to access before being routed to the correct page via the cameraResolver.