CLOUD-10978 - reverted changes to mobile tabs
# Problem
For Downloads, there shouldn't be a mobile tab for Betas and Patches, however the previous Figma design didn't reflect that. The Figma file has now been updated, so the previous changes that were made to add Mobile to Betas and Patches has been reverted. Also the route has been reverted to the correct route for the new download component.

# Solution
Reverted previous changes