CLOUD-11475: Don't display bookmarks that have no device archives
# Problem

Bookmarks with no device archives have no thumbnail/video and are hidden on the VMS. We need to also hide them on Cloud

# Changes

After fetching all the Bookmark data, we'll filter the bookmarks by whether or not they have a device archive and return this new list of Bookmarks

## Notes

I tried to take the list of `deviceIds` that have an archive (`this.system.cameraManager.hasArchives()`) and use this in the `pollParams` for `mediaserver.getBookmarks(pollParams)`

```typescript
pollParams = {
    ...this.bookmarksQuery,
    ...this.buildSearch(),
    deviceId: this.queryParams.deviceId || [], // I'm talking about this
};
```

so that we can just fetch the necessary bookmarks. However, in the case that no Device has an archive, `this.system.cameraManager.hasArchives()` would return an empty array. When setting an empty array for `deviceId` in `pollParams`, this would fetch all bookmarks (including the bookmarks we're trying to hide) which is not what we want, so I'm sticking with this current solution for now where we filter the Bookmarks _after_ they're fetched. If there's a better way to do this, let me know

Related to CLOUD-11475