CLOUD-12398: Improved handling for too many connections
# Problem
When a camera is overloaded with connection requests a wrong error message is shown. This could lead a user to believe that the browser cant play the camera when it's momentarily overwhelmed. A secondary issues is that if the player throws an error it is never cleared so issues like not supported or too many connections becomes stuck for all cameras event though they can play.

# Solution
Added a new error message for too many connections so that the user is informed and knows they can try again shortly. To fix the other issue an effect called clearErrorOnCameraChange was added that resets the player error to an empty string allowing the player to set the error again if it's actually a problem.

(cherry picked from commit 635495e004d60fc7088fac072f6a060b3e1e5442)