CLOUD-12381 CLOUD-12403: Camera permissions bugfix
# Problem

When a user is assigned resource-specific permissions to view/edit a camera (as opposed to being assigned to a built in permission group), sometimes they don't see the camera in the UI when they should. The reason was that camera edit permissions were assigned to `camera.canEdit` on camera initialization, but sometimes this occurred before resource permissions were loaded, and `camera.canEdit` would not stay up to date with any permission updates.



# Solution

Remove the `canEdit` property from the camera object, and instead dynamically check camera permissions wherever they are used using the permissionManager methods. In `settings.component.ts`, this allows camera permissions to come from the up to date `permissions$$` object in the permission manager, rather than the `canEdit` property on the camera which does not stay up to date after initialization.