CLOUD-12415: Added fragment scrolling to Downloads
# Problem
Unless specific build is at the top of the page, when navigating to /download/other/{releaseType}#{buildnumber}, a user couldn't quickly see the release notes or downloads for their build.

# Solution
Add fragment scrolling so that a user is taken to their specific build upon navigating to url with build fragment in route.