CLOUD-11999: Refactored when the login dialog shows up
# Problem
The login dialog was appearing multiple times because multiple functions are triggering it appear. In addition, the loginDialogActive check failed because the login-webadmin.component was closing w/ undefined after each call to open it.

# Solution
Update the login-webadmin.component so that it returns true when it succeeds for a cloud login. The reason this works is that the page will reload when login is successful.

# Changes
- Removed the login-webadmin.service
- Update types for the login-webadmin.component

(cherry picked from commit 7e40f2d82cab51addd33b1ac97839e3052a79f82)