CLOUD-12329 - Initial phase of bookmark sharing landing page
## Problem

For Bookmark Sharing we are creating a page that is not locked behind an auth guard where users can view bookmarks that are shared. This is step one in creating that shared page.

## Solution

* New route at `/share/systemId/bookmarkId`. If you go to `/share` without both of those IDs it will redirect you back home
* A module and component that will do all of the main logic of getting the bookmark and displaying either the bookmark video, a 404, an error, etc.
* A visual only component that will show the bookmark video and some extra bookmark details.

This is not meant to be a final product, but rather an easy to review page to get some early feedback. Feel free to ask questions

![image](/uploads/af40b8a264c05561d8ebf007a550d366/image.png)