CLOUD-12477: Updated boolean check for playerError status
# Problem
If the camera was offline it was considered an error, but because we can play archive that's not always true.

# Solution
Update the check for if the player has an error and only consider offline an error while trying to play the live video feed.

(cherry picked from commit 3d8fc119e2b74048525ed6b80ff771a4929164ac)