CLOUD-11810, CLOUD-11997: Add logout countdown for rejected Terms of Service
# Problem
If a user rejects an updated terms of service, a second dialog appears informing them they will be logged out soon. However, nothing happens after that and they are never logged out.


# Solution
Add a 15 second countdown to the rejected terms of service dialog, and log the user out when that countdown expires.

![image](/uploads/215dc9d35e9d8ee8ec828930dd84d7d3/image.png)