CLOUD-12513: Rename system bugfix
# Problem
If you clear the system name while renaming a system in admin settings, the page goes into a loading state and you can't see the settings again until you refresh the page. This is because we are checking whether `systemName` is truthy as a proxy for the system being fully loaded, and whether to show the admin settings preloader. However, `systemName` can also be set to an empty string when the user renames the system.

# Solution
Narrow the preloader rendering conditional to check whether `systemName` is undefined rather than falsy

(cherry picked from commit 31f0099da90428c4ac20efb354105e4db482567f)