CLOUD-12498: Refactor timeline to allow connecting to timeline state
# Changes
Refactor timeline to be able to accept time periods and cameras from input.

Also split up code into classes by responsibility. The component is responsible for receiving timeline data in a standard format,
TimelineDataModel is responsible for transforming into data useable by the RenderStateModel, and the RenderStateModel is responsible
for rendering.

TODO:

There's still a lot in the RenderStateModel that can probably be cleaned up and there's a few bugs introduced that need to be addressed.

# Reason for the change
Currently the timeline is designed in a way that pretty much requires it to maintain it's own data model, render state, and data fetching.