CLOUD-12477, CLOUD-12478: Fixed issue where playback initialized incorrectly
# Problem
If the camera was offline it still tried to play live when it should have been in the stopped state.

# Solution
When trying to restore playback to live check if the camera is online. If it's not put the player in the stoped state.

(cherry picked from commit 8dc283dacb8d4872c33075b3884a89a2796cfee8)