CLOUD-12587: Fix changing analytics db 5.0
# Problem
Seems like there's a bug on rest/v1/servers when setting metadataStorageId that sometimes causes it to disconnect from cloud and not be able to authenticate cloud users.

# Solution
Reverting to using the old set resource params for 5.0 systems since it still works as expected and would hopefully avoid encountering similar bugs if they exist.

The ServerManager.updateResource and NxSystemAPI.setResourceParams are marked as deprecated and should be removed on the develop branch.

Any uses of ServerManager.updateResource should be split into methods for handling individual properties that can handle differences between mediaserver versions
as more properties get moved out of params.