CLOUD-12575: Fixed issue where systems was blank
# Problem
If there was one 2fa system we didn't initial the page completely resulting in it being blank.

# Solution
In the event that the only system requires 2fa and 2fa is option for the user account we finish the pages initialization.

(cherry picked from commit adb021abf24383ebc70792c488ad4e0187300c1c)