CLOUD-11809: Edit local user permissions bugfix
# Problem

A cloud admin cannot edit a local system user's name/email, when they should be able to. This is because current permissioning logic disables editing of local users that are **not** LDAP users, which seems to be a typo.



# Solution

 Update permissioning logic in the `users-with-roles` component to disable editing of LDAP local users (rather than non-LDAP local users), to match the [corresponding logic](https://gitlab.nxvms.dev/dev/cloud_portal/-/blob/releases_23.3/front_end/libs/features/systems/settings/users/users-with-groups/users-with-groups.component.ts#L77) in the `users-with-groups` component.

![image](/uploads/5645e56aa4b1428b4ca3ee8a98de9e93/image.png){width=70%}

(cherry picked from commit f69c1f5361470e3867008b6f44921db67d5ec3b7)