Fix bug and remove unneeded lines from test_account_server_misc.py
The bug that got me to look at this code is in the lines that
Kyle, Alex and I decided weren't needed. Waiting for the text
about no changes found fixes a bug similar to the original one.
It makes sure that the page has fully loaded. So the function
to check for that was added to the function that checks to see
the page has loaded.