CLOUD-12577, CLOUD-12585: Fixed session renewal in django
# Problem
Sessions weren't renewed correctly in the auto_refresh_token decorator making it impossible to do owner actions. IE disconnect or merge

# Solution
Refactored the auto_refresh_token decorator so that it give priority to the cloud session. Also fixed oauth/logout logic which broke as a part of the auto_refresh_token refactor